+++ packages/orq-rc/src/orq_ai_sdk/agents.py
@@ -3,11 +3,6 @@
 from .basesdk import BaseSDK
 from orq_ai_sdk import models, utils
 from orq_ai_sdk._hooks import HookContext
-from orq_ai_sdk.models import (
-    listagenttasksop as models_listagenttasksop,
-    runagentop as models_runagentop,
-    streamrunagentop as models_streamrunagentop,
-)
 from orq_ai_sdk.types import BaseModel, OptionalNullable, UNSET
 from orq_ai_sdk.utils import eventstreaming, get_security_from_env
 from orq_ai_sdk.utils.unmarshal_json_response import unmarshal_json_response
@@ -394,7 +389,7 @@ class Agents(BaseSDK):
     def retrieve_task(
         self,
         *,
-        id: str,
+        agent_key: str,
         task_id: str,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -405,7 +400,7 @@ class Agents(BaseSDK):
 
         Retrieves detailed information about a specific task for a given agent, including execution status and results.
 
-        :param id:
+        :param agent_key:
         :param task_id:
         :param retries: Override the default retry configuration for this method
         :param server_url: Override the default server URL for this method
@@ -426,13 +421,13 @@ class Agents(BaseSDK):
             base_url = self._get_url(base_url, url_variables)
 
         request = models.GetAgentTaskRequest(
-            id=id,
+            agent_key=agent_key,
             task_id=task_id,
         )
 
         req = self._build_request(
             method="GET",
-            path="/v2/agents/{id}/tasks/{task_id}",
+            path="/v2/agents/{agent_key}/tasks/{task_id}",
             base_url=base_url,
             url_variables=url_variables,
             request=request,
@@ -491,7 +486,7 @@ class Agents(BaseSDK):
     async def retrieve_task_async(
         self,
         *,
-        id: str,
+        agent_key: str,
         task_id: str,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -502,7 +497,7 @@ class Agents(BaseSDK):
 
         Retrieves detailed information about a specific task for a given agent, including execution status and results.
 
-        :param id:
+        :param agent_key:
         :param task_id:
         :param retries: Override the default retry configuration for this method
         :param server_url: Override the default server URL for this method
@@ -523,13 +518,13 @@ class Agents(BaseSDK):
             base_url = self._get_url(base_url, url_variables)
 
         request = models.GetAgentTaskRequest(
-            id=id,
+            agent_key=agent_key,
             task_id=task_id,
         )
 
         req = self._build_request_async(
             method="GET",
-            path="/v2/agents/{id}/tasks/{task_id}",
+            path="/v2/agents/{agent_key}/tasks/{task_id}",
             base_url=base_url,
             url_variables=url_variables,
             request=request,
@@ -589,7 +584,7 @@ class Agents(BaseSDK):
         self,
         *,
         agent_key: str,
-        status: List[models_listagenttasksop.Status],
+        status: List[models.Status],
         limit: Optional[float] = 10,
         starting_after: Optional[str] = None,
         ending_before: Optional[str] = None,
@@ -704,7 +699,7 @@ class Agents(BaseSDK):
         self,
         *,
         agent_key: str,
-        status: List[models_listagenttasksop.Status],
+        status: List[models.Status],
         limit: Optional[float] = 10,
         starting_after: Optional[str] = None,
         ending_before: Optional[str] = None,
@@ -819,10 +814,7 @@ class Agents(BaseSDK):
         self,
         *,
         request: Optional[
-            Union[
-                models_runagentop.RunAgentRequestBody,
-                models_runagentop.RunAgentRequestBodyTypedDict,
-            ]
+            Union[models.RunAgentRequestBody, models.RunAgentRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -915,10 +907,7 @@ class Agents(BaseSDK):
         self,
         *,
         request: Optional[
-            Union[
-                models_runagentop.RunAgentRequestBody,
-                models_runagentop.RunAgentRequestBodyTypedDict,
-            ]
+            Union[models.RunAgentRequestBody, models.RunAgentRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -1012,8 +1001,8 @@ class Agents(BaseSDK):
         *,
         request: Optional[
             Union[
-                models_streamrunagentop.StreamRunAgentRequestBody,
-                models_streamrunagentop.StreamRunAgentRequestBodyTypedDict,
+                models.StreamRunAgentRequestBody,
+                models.StreamRunAgentRequestBodyTypedDict,
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -1125,8 +1114,8 @@ class Agents(BaseSDK):
         *,
         request: Optional[
             Union[
-                models_streamrunagentop.StreamRunAgentRequestBody,
-                models_streamrunagentop.StreamRunAgentRequestBodyTypedDict,
+                models.StreamRunAgentRequestBody,
+                models.StreamRunAgentRequestBodyTypedDict,
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,