+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -79,7 +79,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-10-02T14:54:48.986Z")
+    created: Optional[datetime] = parse_datetime("2025-10-02T15:12:50.746Z")
     r"""The date and time the resource was created"""